Keep trying to connect to ElasticSearch.
If the initial connection fails, just shutting down is not helpful.
Whilst the deployer will continue restarting the service, the aggregate
health check just reports a 404 if the service has stopped rather than
showing the underlying error.

Experimentation reveals that this behaviour is limited to establishing
the initial connection to ElasticSearch; once connected, the connection
can be broken and re-made transparently. Use a channel to inject a
connected client into the service once it becomes available.

Add option to skip tests that require an actual ElasticSearch instance.
See the README for notes on running integration tests.