:sparkles: Add healthcheck that index is writeable.
This check is only part of the /__health endpoint and is severity 2; it
isn't included in the /__gtg, as read requests can still be served. (In
any case it's unlikely that redirecting to another instance of the
service would remedy a read-only problem). The technical summary text
highlights that this condition is expected (and eventually
self-correcting) if a reindexing operation is in progress.