:bug: Fix test cases.
Unfortunately, sending the Elastic client into a channel does not quite behave as per the comment "block until es service has received it", in the sense that the code executing in the for loop at es_service.go line 55 is not guaranteed to execute before the channel sender (es_service_test.go) is unblocked. (This is why the case TestPassClientThroughChannel() waits for up to 1 second before testing that the client has been connected to the service successfully).

Revert most test cases to construct the esService directly (which now requires the construction of a BulkProcessor object too).